Encode and decode using Morse code
Using this code anyone can easily change the language. You can change English to morse code and morse code to English using a Text file. 
you have to upload your text in a text file the name will be this (morse_code.txt) for the English language. and another_Morse_Code.txt for the morse code to English language